link pagination to record to show
