+++ frontend/components/page-contents/feedback-task-content/CenterFeedbackTask.content.vue
@@ -1,24 +1,14 @@
 <template>
-  <div class="wrapper">
-    <RecordFeedbackTaskComponent v-if="record" :record="record" />
-    <QuestionsFormComponent
-      v-if="questionsWithRecordAnswers && questionsWithRecordAnswers.length"
-      :initialInputs="questionsWithRecordAnswers"
-    />
-  </div>
+  <RecordFeedbackTaskAndQuestionnaireContent
+    :datasetId="datasetId"
+    :recordOffset="currentPage"
+    :key="currentPage"
+  />
 </template>
 
 <script>
-import {
-  upsertDatasetQuestions,
-  getQuestionsByDatasetId,
-} from "@/models/feedback-task-model/dataset-question/datasetQuestion.queries";
-import {
-  upsertRecords,
-  getRecordWithFieldsByDatasetId,
-} from "@/models/feedback-task-model/record/record.queries";
-import { getRecordResponsesByRecordId } from "@/models/feedback-task-model/record-response/recordResponse.queries";
-
+import { upsertDatasetQuestions } from "@/models/feedback-task-model/dataset-question/datasetQuestion.queries";
+import { upsertRecords } from "@/models/feedback-task-model/record/record.queries";
 export default {
   name: "CenterFeedbackTaskContent",
   props: {
@@ -26,16 +16,10 @@ export default {
       type: String,
       required: true,
     },
-    orderBy: {
-      type: Object,
-      default: () => {
-        return { orderQuestionBy: "order", ascendent: true };
-      },
-    },
   },
   data() {
     return {
-      recordOffset: 0,
+      currentPage: 0,
     };
   },
   created() {
@@ -248,7 +232,7 @@ export default {
       },
       {
         id: "id_7",
-         name:"nameOfQuestion_id_7",
+        name: "nameOfQuestion_id_7",
         question: "Comment",
         placeholder: "this is the placeholder",
         options: [
@@ -273,67 +257,14 @@ export default {
     upsertRecords(formattedRecords);
     upsertDatasetQuestions(formattedQuestions);
 
-    // RETRIEVES records from ORM
-    this.record = getRecordWithFieldsByDatasetId(
-      this.datasetId,
-      1,
-      this.recordOffset
-    );
-
-    this.questionsWithRecordAnswers = null;
-    if (this.record) {
-      this.recordResponses = getRecordResponsesByRecordId(this.record.id);
-
-      // COMPUTE QUESTIONS
-      // INIT questions with record response answers
-      this.questions = getQuestionsByDatasetId(
-        this.datasetId,
-        this.orderBy?.orderQuestionsBy,
-        this.orderBy?.ascendent
-      );
-
-      // COMPUTE questions with responses from record
-      const newOptionsByQuestion = this.factoryNewOptionsByQuestion();
-      this.questionsWithRecordAnswers =
-        this.factoryQuestionsWithRecordAnswer(newOptionsByQuestion);
-    }
+    this.onBusEventIsLoadingLabels()
   },
   methods: {
-    factoryQuestionsWithRecordAnswer(newOptionsByQuestion) {
-      const questionsWithRecordAnswers = this.questions.map((question) => {
-        const newOptions =
-          newOptionsByQuestion.find(
-            (recordResponseByQuestion) =>
-              recordResponseByQuestion.question_id === question.id
-          )?.newOptions || question.options;
-        return { ...question, options: newOptions };
+    onBusEventIsLoadingLabels() {
+      this.$root.$on("current-page", (currentPage) => {
+        //NOTE - the pagination start at 1 but the record start at 1 => there is an offset of 1 to remove 
+        this.currentPage = currentPage-1;
       });
-
-      return questionsWithRecordAnswers;
-    },
-    factoryNewOptionsByQuestion() {
-      const newOptionsByQuestion = [];
-      this.questions.forEach((question) => {
-        this.recordResponses.forEach((response) => {
-          if (response.question_id === question.id) {
-            const newOptions = question.options.map((output) => {
-              const recordResponseOutputWithSameTextAsInQuestionOutput =
-                response.options.find(
-                  (responseOutput) => responseOutput.id === output.id
-                );
-
-              if (recordResponseOutputWithSameTextAsInQuestionOutput) {
-                return recordResponseOutputWithSameTextAsInQuestionOutput;
-              }
-              return output;
-            });
-
-            newOptionsByQuestion.push({ question_id: question.id, newOptions });
-          }
-        });
-      });
-
-      return newOptionsByQuestion;
     },
     factoryRecordsForOrm(records) {
       return records.map((record) => {
@@ -360,13 +291,8 @@ export default {
       });
     },
   },
+  destroyed() {
+    this.$root.$off("current-page");
+  },
 };
 </script>
-
-<style lang="scss" scoped>
-.wrapper {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 2 * $base-space;
-}
-</style>
